modules/dashboard_requirements.txt
# Core Streamlit and database dependencies
streamlit>=1.28.0
pandas>=1.5.0
psycopg2-binary>=2.9.0

# Visualization libraries
plotly>=5.15.0

# Configuration management
PyYAML>=6.0

# Additional utilities
python-dateutil>=2.8.0
